microservice_ecommerce/admin/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
mysqlclient==2.2.4
django-mysql==4.12.0
django-cors-headers==4.3.1
pika==1.3.2
pillow==10.3.0